Added some fixes and changes to the functions, moved email verfification to auth route
